% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_calculate_counts.R
\name{ww_apply_cole_kripke}
\alias{ww_apply_cole_kripke}
\title{Process Count Data}
\usage{
ww_apply_cole_kripke(data)
}
\arguments{
\item{data}{A `data.frame` from `acti_calculate_counts` that has
columns `axis1-3` and `counts`}
}
\value{
A `data.frame` of transformed data

A `data.frame` of transformed data
}
\description{
Process Count Data

Process Count Data
}
\note{
This calls the downstream wear-processing helpers used by
`actigraph.sleepr`
}
\examples{
\dontrun{
path = actiread::acti_example_gt3x()
ac = actiread::acti_read_gt3x(path)
out = acti_calculate_counts(ac)
}
data = actimetrics::acti_count_data
wear = actimetrics::acti_calculate_wear(data)
tro_wear = actimetrics::acti_calculate_wear(data, method = "troiano")
ck = actimetrics::acti_apply_cole_kripke(data)
tl = actimetrics::acti_apply_tudor_locke(ck)
sadeh = actimetrics::acti_apply_sadeh(ck)
}
