% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_calculate_counts.R
\name{ww_calculate_counts}
\alias{ww_calculate_counts}
\title{Process Count Data}
\usage{
ww_calculate_counts(
  data,
  epoch = 60L,
  resample = TRUE,
  lfe_select = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A `data.frame` from `acti_calculate_counts` that has
columns `axis1-3` and `counts`}

\item{epoch}{epoch length in seconds.  Default is 60 seconds.
See `agcounts::calculate_counts()`}

\item{resample}{(recommended) resample the data to 30Hz using
[actibase::resample] vs. using the resampling method from
[agcounts::calculate_counts].}

\item{lfe_select}{Apply the Actigraph Low Frequency Extension filter.
See `agcounts::calculate_counts()`
higher values are higher levels of verbosity.}

\item{verbose}{print diagnostic messages.  Either logical or integer, where}
}
\value{
A `data.frame` of transformed data

A `data.frame` of transformed data
}
\description{
Process Count Data

Process Count Data
}
\note{
This calls the downstream wear-processing helpers used by
`actigraph.sleepr`
}
\examples{
path = ww_example_gt3x_file()
ac = ww_read_gt3x(path, verbose = FALSE)
out = ww_calculate_counts(ac)
}
