% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_calculate_counts.R
\name{ww_calculate_wear}
\alias{ww_calculate_wear}
\alias{ww_calculate_nonwear}
\title{Process Count Data}
\usage{
ww_calculate_wear(
  data,
  method = c("choi", "troiano"),
  use_magnitude = TRUE,
  ...
)

ww_calculate_nonwear(
  data,
  method = c("choi", "troiano"),
  use_magnitude = TRUE,
  ...
)
}
\arguments{
\item{data}{A `data.frame` from `acti_calculate_counts` that has
columns `axis1-3` and `counts`}

\item{method}{Method for detecting non-wear, either "choi" or "troiano",
corresponding to `actigraph.sleepr::apply_choi()` or
`actigraph.sleepr::apply_troiano()`}

\item{use_magnitude}{If `TRUE`, the magnitude of the vector
(axis1, axis2, axis3) is used to measure activity;
otherwise the axis1 value is used.}

\item{...}{additional arguments to pass to `actigraph.sleepr` function}
}
\value{
A `data.frame` of transformed data

A `data.frame` of transformed data
}
\description{
Process Count Data

Process Count Data
}
\note{
This calls the downstream wear-processing helpers used by
`actigraph.sleepr`
}
\examples{
\dontrun{
path = actiread::acti_example_gt3x()
ac = actiread::acti_read_gt3x(path)
out = acti_calculate_counts(ac)
}
data = actimetrics::acti_count_data
wear = actimetrics::acti_calculate_wear(data)
tro_wear = actimetrics::acti_calculate_wear(data, method = "troiano")
ck = actimetrics::acti_apply_cole_kripke(data)
tl = actimetrics::acti_apply_tudor_locke(ck)
sadeh = actimetrics::acti_apply_sadeh(ck)
}
