% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_epa_walkability.R
\name{ww_epa_walkability}
\alias{ww_epa_walkability}
\title{Get EPA Walkability Index}
\usage{
ww_epa_walkability(geoid, geometry = TRUE, ...)
}
\arguments{
\item{geoid}{GEOID10 of the area of interest.  This should be a 12 character string.
If \code{NULL}, then all GEOIDs are selected (warning: this can be a lot of data, take a while,
use with caution).}

\item{geometry}{Should geometry be returned?  Passed to \code{\link[arcgislayers:arc_select]{arcgislayers::arc_select()}}}

\item{...}{additional arguments to pass to \code{\link[arcgislayers:arc_select]{arcgislayers::arc_select()}}}
}
\value{
A \code{data.frame} of results
}
\description{
Get EPA Walkability Index
}
\note{
See
\url{https://geodata.epa.gov/arcgis/rest/services/OA/WalkabilityIndex/MapServer/0}
}
\examples{
\dontshow{if (rlang::is_installed("arcgislayers")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ww_epa_walkability(c("240054519002", "240054026041", "245102303002"))
\dontshow{\}) # examplesIf}
}
