% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_example_data.R
\name{ww_example_sensorlog_file}
\alias{ww_example_sensorlog_file}
\alias{ww_example_gt3x_file}
\alias{ww_example_sensorlogger_file}
\alias{ww_example_sensorlogger_location_file}
\title{Activity/Actigraphy Example Data}
\usage{
ww_example_sensorlog_file()

ww_example_gt3x_file()

ww_example_sensorlogger_file()

ww_example_sensorlogger_location_file()
}
\value{
A file path
}
\description{
Activity/Actigraphy Example Data
}
\examples{
library(actiread)
acti_example_gt3x()
acti_example_cwa()
acti_example_sensorlogger_file()
}
