% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_fips.R
\name{ww_fips15}
\alias{ww_fips15}
\alias{ww_fips12}
\title{Get Census Tract FIPS codes}
\usage{
ww_fips15(state, county, tract, block = NA)

ww_fips12(state, county, tract, block = NA)
}
\arguments{
\item{state}{2-digit state FIPS code}

\item{county}{3-digit state FIPS code}

\item{tract}{6-digit tract FIPS code}

\item{block}{6-digit block FIPS code.  If omitted or \code{NA}, 12-digit codes returned}
}
\value{
A 12-to-15 digit FIPS code
}
\description{
Get Census Tract FIPS codes
}
\examples{
ww_fips15(24, 510, 60400)
ww_fips15(24, 510, 60400, block = 2002)
ww_fips12(24, 510, 60400, block = 2002)
}
