% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_read_gt3x.R
\name{ww_read_gt3x}
\alias{ww_read_gt3x}
\title{Read GT3X file}
\usage{
ww_read_gt3x(
  path,
  asDataFrame = TRUE,
  imputeZeroes = TRUE,
  verbose = TRUE,
  ...,
  fill_zeroes = TRUE,
  apply_tz = FALSE,
  check_attributes = TRUE,
  tz = "GMT"
)
}
\arguments{
\item{path}{Path to gt3x file}

\item{asDataFrame}{convert to an `activity_df`, see
\code{as.data.frame.activity}}

\item{imputeZeroes}{Impute zeros in case there are missingness?
Default is `FALSE`, in which case
the time series will be incomplete in case there is missingness.}

\item{verbose}{print diagnostic messages, higher values = more verbosity.}

\item{...}{additional arguments to pass to [read.gt3x::read.gt3x()]}

\item{fill_zeroes}{Rows with all zeros will be filled in with the last
observation carried forward as is done with ActiLife.  Recommended}

\item{apply_tz}{Apply the timezone from the header `TimeZone` attribute,
if available}

\item{check_attributes}{Check that the attributes are included This is a sanity check,
including checking that `sample_rate` is in the attributes.}

\item{tz}{timezone to project the data into.  The data read in via
[read.gt3x::read.gt3x()] says the timezone is GMT, but the time values is in the
native timezone.  So this data is projected into the correct time zone and then
forced into the timezone given by `tz`.  Set to `NULL` to not apply this
forcing.}
}
\value{
A `data.frame`
}
\description{
Read GT3X file
}
\examples{
path = ww_example_gt3x_file()
ac = ww_read_gt3x(path, verbose = FALSE)
}
