% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_read_sensorlogger_zzz.R
\name{ww_read_sensorlogger}
\alias{ww_read_sensorlogger}
\title{Read SensorLogger Data}
\usage{
ww_read_sensorlogger(file, verbose = FALSE, ...)
}
\arguments{
\item{file}{A character vector of SensorLogger files, usually from unzipping
the file, or a zip file of SensorLogger files}

\item{verbose}{print diagnostic messages.  Either logical or integer, where
higher values are higher levels of verbosity.}

\item{...}{additional arguments to pass to \code{\link[readr:read_csv]{readr::read_csv()}}.
If \code{verbose = FALSE}, then \code{progress = FALSE} and \code{show_col_types = FALSE},
unless otherwise overridden}
}
\value{
A \code{data.frame} of data
}
\description{
Read SensorLogger Data
}
