% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_read_sensorlogger_simple.R
\name{ww_read_sensorlogger_general}
\alias{ww_read_sensorlogger_general}
\title{Read SensorLogger Data}
\usage{
ww_read_sensorlogger_general(file, ..., verbose = FALSE)
}
\arguments{
\item{file}{A character vector of SensorLogger files, usually from unzipping
the file, or a zip file of SensorLogger files}

\item{...}{additional arguments to pass to [readr::read_csv()].
If `verbose = FALSE`, then `progress = FALSE` and `show_col_types = FALSE`,
unless otherwise overridden}

\item{verbose}{print diagnostic messages.  Either logical or integer, where
higher values are higher levels of verbosity.}
}
\value{
A `data.frame` of data
}
\description{
Read SensorLogger Data
}
