% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_read_sensorlogger_location.R
\name{ww_sensorlogger_location_colnames_mapping}
\alias{ww_sensorlogger_location_colnames_mapping}
\alias{ww_sensorlogger_location_spec}
\alias{ww_read_sensorlogger_location}
\title{Read SensorLogger Data}
\usage{
ww_sensorlogger_location_colnames_mapping()

ww_sensorlogger_location_spec()

ww_read_sensorlogger_location(file, ...)
}
\arguments{
\item{file}{A character vector of SensorLogger files, usually from unzipping
the file, or a zip file of SensorLogger files}

\item{...}{additional arguments to pass to [readr::read_csv()].
If `verbose = FALSE`, then `progress = FALSE` and `show_col_types = FALSE`,
unless otherwise overridden}
}
\value{
A `data.frame` of data
}
\description{
Read SensorLogger Data
}
