% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_separate_times.R
\name{ww_separate_times}
\alias{ww_separate_times}
\title{Separate Times into Date, Hour, and Minute}
\usage{
ww_separate_times(data)
}
\arguments{
\item{data}{a `data.frame` with a `time` column}
}
\value{
A `data.frame` with date, hour, minute, and day columns
}
\description{
Separate Times into Date, Hour, and Minute
}
\examples{
acti_separate_time(acti_raw_data)
acti_create_date(acti_raw_data)
acti_create_hour(acti_raw_data)
acti_create_minute(acti_raw_data)
}
