% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_summarise_sensorlog.R
\name{ww_summarize_sensorlog}
\alias{ww_summarize_sensorlog}
\alias{ww_summarise_sensorlog}
\alias{ww_minute_sensorlog}
\alias{ww_summarize_distance_sensorlog}
\title{Summarize SensorLog Data}
\usage{
ww_summarize_sensorlog(data)

ww_summarise_sensorlog(data)

ww_minute_sensorlog(data, seconds = 60L)

ww_summarize_distance_sensorlog(data)
}
\arguments{
\item{data}{A SensorLog-style `data.frame`, usually output from
[acti_process_sensorlog]}

\item{seconds}{integer of the number of seconds to summarize the data
for the "minute" level. Usually 1 minute/60 seconds.  For
\code{ww_summarize_distance_sensorlog}, summarization is done depending on how the
data is grouped.}
}
\value{
The `data.frame` with the summarized data for each date
}
\description{
Summarize SensorLog Data
}
