% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_calculate_counts.R
\name{ww_calculate_counts}
\alias{ww_calculate_counts}
\title{Process SensorLog Daa}
\usage{
ww_calculate_counts(data, epoch = 60L, lfe_select = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} from \link{ww_read_gt3x}}

\item{verbose}{print diagnostic messages.  Either logical or integer, where
higher values are higher levels of verbosity.}

\item{...}{additional arguments to pass to \link{ww_process_time}}
}
\value{
A \code{data.frame} of transformed data
}
\description{
Process SensorLog Daa
}
\note{
This calls \link{ww_check_data}, \link{ww_calculate_distance}, and
\link{ww_process_time}
}
\examples{
path = ww_example_gt3x_file()
ac = ww_read_gt3x(path, verbose = FALSE)
out = ww_calculate_counts(ac)
}
