% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_example_data.R
\name{ww_example_sensorlog_file}
\alias{ww_example_sensorlog_file}
\alias{ww_example_gt3x_file}
\title{Paths to Example data}
\usage{
ww_example_sensorlog_file()

ww_example_gt3x_file()
}
\value{
A character file name
}
\description{
Simple wrapper of \code{\link[base:system.file]{base::system.file()}} to get the files included in
the package
}
\examples{
ww_example_sensorlog_file()
ww_example_gt3x_file()
}
