% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_read_gt3x.R
\name{ww_info_gt3x}
\alias{ww_info_gt3x}
\title{Read GT3X file}
\usage{
ww_info_gt3x(path, ...)
}
\arguments{
\item{path}{Path to gt3x file}

\item{...}{additional arguments to pass to [read.gt3x::read.gt3x()]}
}
\value{
A `data.frame`
}
\description{
Read GT3X file
}
\examples{
data = acti_read_gt3x(acti_example_gt3x())
data = acti_read_gt3x(
  acti_example_gt3x(),
  tz = NULL,
  apply_tz = FALSE,
  verbose = FALSE,
  fill_zeroes = FALSE
)
info = acti_info_gt3x(acti_example_gt3x())
}
