% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_calculate_counts.R
\name{ww_process_gt3x}
\alias{ww_process_gt3x}
\title{Process Count Data}
\usage{
ww_process_gt3x(
  data,
  lfe_select = FALSE,
  method = c("choi", "troiano"),
  use_magnitude = TRUE,
  verbose = TRUE,
  ...,
  apply_tz = TRUE
)
}
\arguments{
\item{data}{A `data.frame` from `acti_calculate_counts` that has
columns `axis1-3` and `counts`}

\item{lfe_select}{Apply the Actigraph Low Frequency Extension filter.
See `agcounts::calculate_counts()`
higher values are higher levels of verbosity.}

\item{method}{Method for detecting non-wear, either "choi" or "troiano",
corresponding to `actigraph.sleepr::apply_choi()` or
`actigraph.sleepr::apply_troiano()`}

\item{use_magnitude}{If `TRUE`, the magnitude of the vector
(axis1, axis2, axis3) is used to measure activity;
otherwise the axis1 value is used.}

\item{verbose}{print diagnostic messages.  Either logical or integer, where}

\item{...}{additional arguments to pass to `actigraph.sleepr` function}

\item{apply_tz}{apply timezone to \link[actiread:acti_read_gt3x]{actiread::acti_read_gt3x}.}
}
\description{
Process Count Data
}
\note{
For `acti_process_gt3x`, the `...` argument are passed to
`actiread::acti_read_gt3x()`
}
