% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_process_sensorlog.R
\name{ww_process_sensorlog}
\alias{ww_process_sensorlog}
\alias{ww_check_data}
\alias{ww_calculate_distance}
\title{Process SensorLog Daa}
\usage{
ww_process_sensorlog(
  data,
  lat = NULL,
  lon = NULL,
  dist_fun = geosphere::distVincentyEllipsoid,
  expected_timezone = "America/New_York"
)

ww_check_data(data)

ww_calculate_distance(
  data,
  lat,
  lon,
  dist_fun = geosphere::distVincentyEllipsoid
)
}
\arguments{
\item{data}{A \code{data.frame} from \link{ww_read_sensorlog}}

\item{lat}{Latitude of central point (e.g. home) to calculate distance.
Set to \code{NULL} if distnace not to be run.}

\item{lon}{Longitude of central point (e.g. home) to calculate distance
Set to \code{NULL} if distnace not to be run.}

\item{dist_fun}{Distance function to pass to \link[geosphere:distm]{geosphere::distm}}

\item{expected_timezone}{Expected Timezone based on the latitude/longitude
of the data based on the lat/lon values from SensorLog.  Set to
\code{NULL} if not to be checked.}
}
\value{
A \code{data.frame} of transformed data
}
\description{
Process SensorLog Daa
}
