% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_process_time.R
\name{ww_process_time}
\alias{ww_process_time}
\title{Process the Time data from SensorLog and Compare to an Expected Timezone}
\usage{
ww_process_time(data, expected_timezone = "America/New_York")
}
\arguments{
\item{data}{A \code{data.frame} from \link{ww_read_sensorlog}}

\item{expected_timezone}{Expected Timezone based on the latitude/longitude
of the data based on the lat/lon values from SensorLog (
e.g. \code{"America/New_York"}).  Set to
\code{NULL} if not to be checked.}
}
\value{
A \code{data.frame}
}
\description{
Process the Time data from SensorLog and Compare to an Expected Timezone
}
