% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_process_time.R
\name{ww_process_time}
\alias{ww_process_time}
\title{Process the Time data from SensorLog and Compare to an Expected Timezone}
\usage{
ww_process_time(data, expected_timezone = "America/New_York", tz = "GMT", ...)
}
\arguments{
\item{tz}{timezone to project the data into.  Keeping as \code{GMT} to agree
with ActiGraph, passed to \link[lubridate:as_date]{lubridate::as_datetime}.}

\item{...}{additional arguments to pass to \code{\link[lutz:tz_lookup_coords]{lutz::tz_lookup_coords()}}}
}
\value{
A \code{data.frame}
}
\description{
Process the Time data from SensorLog and Compare to an Expected Timezone
}
