% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_read_sensorlog.R
\name{ww_read_sensorlog}
\alias{ww_read_sensorlog}
\alias{ww_csv_spec}
\alias{ww_sensorlog_csv_colnames_mapping}
\title{Read SensorLog Data}
\usage{
ww_read_sensorlog(files, verbose = FALSE, robust = FALSE)

ww_csv_spec()

ww_sensorlog_csv_colnames_mapping()
}
\arguments{
\item{files}{A character vector of SensorLog files, usually from unzipping
the file}

\item{verbose}{print diagnostic messages.  Either logical or integer, where
higher values are higher levels of verbosity.}

\item{robust}{if \code{TRUE} then \link{rewrite_sensorlog_csv} is run on the data
to try to fix any shifts with the data.}
}
\value{
A \code{data.frame} of data
}
\description{
Read SensorLog Data
}
