% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_read_sensorlogger_location.R,
%   R/ww_read_sensorlogger_zzz.R
\name{ww_sensorlogger_location_colnames_mapping}
\alias{ww_sensorlogger_location_colnames_mapping}
\alias{ww_sensorlogger_location_spec}
\alias{ww_read_sensorlogger_location}
\alias{ww_read_sensorlogger}
\alias{ww_read_sensorlogger_general}
\alias{ww_read_sensorlogger_accelerometer}
\alias{ww_read_sensorlogger_accelerometer_uncalibrated}
\alias{ww_read_sensorlogger_annotation}
\alias{ww_read_sensorlogger_battery}
\alias{ww_read_sensorlogger_gravity}
\alias{ww_read_sensorlogger_gyroscope_uncalibrated}
\alias{ww_read_sensorlogger_metadata}
\alias{ww_read_sensorlogger_orientation}
\alias{ww_read_sensorlogger_pedometer}
\title{Read SensorLogger Data}
\usage{
ww_sensorlogger_location_colnames_mapping()

ww_sensorlogger_location_spec()

ww_read_sensorlogger_location(file, ...)

ww_read_sensorlogger(file, verbose = FALSE, ...)

ww_read_sensorlogger_general(file, ..., verbose = FALSE)

ww_read_sensorlogger_accelerometer(file, ..., verbose = FALSE)

ww_read_sensorlogger_accelerometer_uncalibrated(file, ..., verbose = FALSE)

ww_read_sensorlogger_annotation(file, ..., verbose = FALSE)

ww_read_sensorlogger_battery(file, ..., verbose = FALSE)

ww_read_sensorlogger_gravity(file, ..., verbose = FALSE)

ww_read_sensorlogger_gyroscope_uncalibrated(file, ..., verbose = FALSE)

ww_read_sensorlogger_metadata(file, ..., verbose = FALSE)

ww_read_sensorlogger_orientation(file, ..., verbose = FALSE)

ww_read_sensorlogger_pedometer(file, ..., verbose = FALSE)
}
\arguments{
\item{file}{A character vector of SensorLogger files, usually from unzipping
the file, or a zip file of SensorLogger files}

\item{...}{additional arguments to pass to \code{\link[readr:read_csv]{readr::read_csv()}}.
If \code{verbose = FALSE}, then \code{progress = FALSE} and \code{show_col_types = FALSE},
unless otherwise overridden}

\item{verbose}{print diagnostic messages.  Either logical or integer, where
higher values are higher levels of verbosity.}
}
\value{
A \code{data.frame} of data
}
\description{
Read SensorLogger Data
}
