% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ww_summarise_sensorlog.R
\name{ww_summarize_sensorlog}
\alias{ww_summarize_sensorlog}
\alias{ww_summarise_sensorlog}
\title{Summarize SensorLog Data}
\usage{
ww_summarize_sensorlog(data, seconds = 60L)

ww_summarise_sensorlog(data, seconds = 60L)
}
\arguments{
\item{data}{\code{data.frame} of the data, output from \code{ww_process_sensorlog}}

\item{seconds}{integer of the number of seconds to summarize the data,
usually 1 minute}
}
\value{
The \code{data.frame} with the summarized data, by taking the
mean of
}
\description{
Summarize SensorLog Data
}
